- fixes #7:
  - deleted the less feature rich FinalResults/multiDimFit.py

  - moved ResultScripts/multiDimFit.py to FinalResults/multiDimFit.py

  - softlinked FinalResults/multiDimFit.py to ResultScripts/multiDimFit.py

  so the actual file is now in FinalResults and ResultScripts
  has a softlink
